Remove .npmrc and update dependencies
The .npmrc file has been deleted as it is no longer needed, and references to it in the Dockerfile have been commented out. Updated various dependencies in `package.json` and the lock file to their latest stable versions for improvements and fixes. This includes updates to TypeScript-related packages and other utility libraries.